+++ simple_deploy/management/commands/heroku/platform_deployer.py
@@ -209,10 +209,12 @@ class PlatformDeployer:
 
         if path.exists():
             self.sd.write_output("    Found existing Procfile.")
-            proceed = self.sd.get_confirmation(self.messages.procfile_found)
+            # proceed = self.sd.get_confirmation(self.messages.procfile_found)
+            proceed = self.sd.messages.file_found("Procfile")
             if not proceed:
                 raise self.sd.utils.SimpleDeployCommandError(
-                    self.messages.cant_overwrite_procfile
+                    # self.messages.cant_overwrite_procfile
+                    self.sd.messages.file_replace_rejected
                 )
 
         # No Procfile exists, or we're free to write over existing one.